requirements.txt
streamlit==1.32.2
beautifulsoup4==4.12.3
requests==2.31.0
pillow==10.2.0
python-dotenv==1.0.1